make chatwork module independent
